feat: on server side permissions
