+++ src/services/ServerRequest.ts
@@ -1,0 +1,1 @@
+export const serverRequest = async (url: string) => {};
